chore: moved initial value derivation back to Preferences class, outside of Storage
